Add support for "template" types (https://redirect.github.com/duckdb/duckdb/pull/18410)
This PR adds basic support for defining functions with "template" types.

Template types function like "placeholder" types when defining a
function, similarly to the already existing `ANY` type. The main
difference is that instead of relying on the function to implement a
custom "bind" callback in order to replace the `ANY` type with a
concrete type after inspecting the function arguments, `TEMPLATE` types
are automatically "inferred" to the most general type of all types
passed to them, and replaced by the binder, before the function
callbacks are invoked. Additionally, it's possible to declare a function
with multiple _different_ `TEMPLATE` types by declaring them with an
"id".

## Consistent inference logic and errors

In particular, functions that operate on nested types have previously
required complex binding logic, and generally struggle with handling
`SQLNULL` and `UNKNOWN` (from prepared statements) properly. By using
template types, all logic to deal with inferring argument/return types
is now consistent and centralized in one location.

The actual inference logic is pretty simple and consist of two steps,
__Inference__ and __Substitution__.

In the __Inference__ phase, we recursively visit the type in a function
parameter simultaneously as the type of the corresponding argument, and
compare any nested types "positionally'. Once we find a template (child)
type on the parameters side, we "bind" the template "id" to the
corresponding (child) type on the argument side. If the template was
already bound, and the bound type differs from the type on the argument
side, we try to "promote" the bound type with the new type using
`MaxLogicalType` to "widen" the inferred template type. If the promotion
fails, we throw an error explaining the inference steps so far, which is
now consistent across functions and much more detailed, for example:

```sql
select list_value(1, 1, 1, '1', 1.10, [2]);
----
Binder Error:
Cannot deduce template type 'T1' in function: 'list_value(T1, [T1...]) -> T1[]'
Type 'T1' was inferred to be:
 - 'INTEGER_LITERAL', from first occurence
 - 'INTEGER', by promoting 'INTEGER_LITERAL' + 'STRING_LITERAL'
 - 'DECIMAL(12,2)', by promoting 'INTEGER' + 'DECIMAL(3,2)'
 - 'INTEGER[]', which is incompatible with previously inferred type!

LINE 1: select list_value(1, 1, 1, '1', 1.10, [2]);
                                               ^
```

Once we've tried to bind all template types, we call the functions
"bind" callback (if present) to give the callback a chance to resolve
any unbound templated, if they would like to.

Then, in the __Substitution__ phase, we again iterate through and visit
all function parameter types, but also the function "varargs" type, and
return type, and now actually replace the template types with their
inferred types. If we find a template which does not have a type
inferred for it, we also throw an error.

### Better function signatures

The printed signatures of generic functions are much more informative,
as it's now possible to actually distinguish the relationship between
different generic (templated) parameters, and sometimes even possible to
identify the behavior of certain edge cases (e.g. `list_value()`).
Here's some examples:

- `map_contains(ANY, ANY) -> ANY` => `map_contains(MAP(T1, T2), T1) ->
T1`
- `list_value(ANY) -> LIST` => `list_value(T1, [T1...]) -> T1[]` and
`list_value() -> NULL[]`

### Behavior changes

By replacing the custom binding logic in a lot of functions with
template types instead, I've found that some of the old behavior was
inconsistent in a way that I would consider buggy. I've thus made some
adjustments.

- There is now a separate overland for the "empty" `MAP()` case, instead
of trying to change the behavior in the bind by inspecting if there are
0 or 2 arguments.
- `list_value()` now has a separate overload for the "empty list" case
when no arguments are passed.
- Lots of map functions .e.g `map_keys`, `map_values`, `map_extract`
used to be defined as taking varargs, but then check and error in the
bind callback if there are more than the expected amount of arguments.
This has now changed so that they are no longer varargs and have well
defined signatures
- `map_extract` used to return `[]` when passed `SQLNULL`, but `NULL`
when passed any typed null. This is has now changed to always return
null regardless if passed a typed or untyped null.
- `map_extract` used to accept anything as the key lookup argument, even
if the type didn't match the actual type of the key, and then return
`NULL`. This now throws a proper type error.
- `array_agg()`/`list()` used to set their return type to `SQLNULL` when
passed a prepared statement parameter. Since they are now defined with
template types they now properly throw `ParameterNotResolved`
exceptions.

### Limitations

Only templated child types in "unnamed" `STRUCT` can be inferred for
now. `UNION` types are also not supported yet. This doesn't mean you
can't pass a `STRUCT` or `UNION` to a template - you just can't declare
a function with a signature like `foo(STRUCT(x T1, i INT), arg INT) ->
T1`. DuckDB does not have any of these in core, so it's a bit difficult
to test, but we want to enable this so extensions can define functions
like this in a follow-up PR.

I also haven't changed all of our nested functions that use `ANY` to use
`TEMPLATE` yet, as some are very advanced and I don't want to risk
breaking them yet.

### Future work

There's lots more I want to do with this in the future:

- Support templated __non-type type-parameters__ as well, so that you
can define e.g .`array_concat(INT[N1], INT[N2]) -> INT[N1 + N2]` or
`decimal_add(DECIMAL(W1, S1), DECIMAL(W1, S1)) -> DECIMAL(W1, S1)`

- Support for typing __lambdas__, e.g. `list_select(T1[], (T1 ->
BOOLEAN)) -> T1[]`

- Support for __type constraints/traits__, e.g. `list_sort(T1[]) where
T1: Sortable` or `safe_add(T1, T1) -> T1 where T1: Numeric`

Upstream-commit: https://github.com/duckdb/duckdb/commit/2b974a241bfed0c3d823286445595419a3b599ce