Change ctrl-a/ctrl-e to move to start/end of line, not buffer (https://redirect.github.com/duckdb/duckdb/pull/18490)
This cli behaviour has been driving me crazy, but perhaps it's only me.
I think that for a multi-line query, you should go to the start/end of
the current line with ctrl-a/ctrl-e, not start/end of the entire
multi-line buffer. If you want to go to the start/end of the buffer, the
home/end keys are bound to that.

It could be a personal preference of mine, or maybe I spend too much
time in the cli, but I think this is more intuitive behaviour for
multi-line queries as it mirrors how multi-line editing works in some
other apps, e.g web browser, word processors etc.

Upstream-commit: https://github.com/duckdb/duckdb/commit/277a91de3e91e7c1a660edb6e3664616c93b9633