Fix: Remove overly strict assertion on empty string value (https://redirect.github.com/duckdb/duckdb/pull/18504)
Fixes https://github.com/duckdblabs/duckdb-internal/issues/5475

This removes the assertion `D_ASSERT(!entry_value.empty())`.

This assertion was based on the assumption that an empty string was an
invalid state at this point. However, I think empty strings are a valid
case and are handled gracefully by downstream code. The assertion was
therefore too aggressive and has been removed.

Upstream-commit: https://github.com/duckdb/duckdb/commit/5633a979536d58221ca2643d46ef5b98165dd650