Dynamically determine dictionary size limit in Parquet writer (if unset) (https://redirect.github.com/duckdb/duckdb/pull/18356)
It is currently static and tied to `ROW_GROUP_SIZE` if unset, so if we
write smaller row groups (due to less data or `ROW_GROUP_SIZE_BYTES`),
the dictionary can be as big as the row group.

This PR makes the dictionary size dynamic based on the _actual row group
size_. This speeds up some edge-case Parquet writes (e.g., with large
`ROW_GROUP_SIZE` and small `ROW_GROUP_SIZE_BYTES`) and improves the
compression ratio (by bailing on dictionary when it is not fruitful).

If the dictionary size limit is set manually, the behavior is the same
as before.

Upstream-commit: https://github.com/duckdb/duckdb/commit/28733bfdddaa7760d8ef73b2611e2015c07bfbe2