Unittester: Add the `--sort-style` parameter that allows a fallback comparison where results are sorted according to a given sort-style (https://redirect.github.com/duckdb/duckdb/pull/18542)
This PR adds the `sort_style` parameter that allows a fallback
comparison for query results in the tester, e.g. `--sort-style rowsort`.
In normal operation the unittester is order-sensitive. However, several
of the DuckDB tests are written with certain assumptions about order
(e.g. that DuckDB is by default insertion-order preserving). This is
correct for DuckDB files, but is not true for all catalog back-ends,
e.g. DuckLake or Postgres are not insertion-order preserving. In order
to allow for more tests to be re-run with other catalogs, we add the
option to add a fallback sort style for results which considers results
equivalent using order insensitive comparisons.

Upstream-commit: https://github.com/duckdb/duckdb/commit/36423ca446414166a3d596794e719efb74187956