Force `LIST`/`ARRAY` child vectors on a Parquet single page (https://redirect.github.com/duckdb/duckdb/pull/18578)
Fixes https://redirect.github.com/duckdb/duckdb/pull/18512

DuckDB can read/write Parquet lists that are spread out over multiple
pages, but this is apparently not what the Parquet spec intended. This
PR forces DuckDB Parquet writes to keep lists on a single Parquet page.
It's not very fine-grained, as it forces batches of
`STANDARD_VECTOR_SIZE` lists to be on a single Parquet page, but I doubt
that will be an issue, and this is the easiest way to comply with the
spec.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6fd7f060b19ca2379980a8e6e36469ee5de99fa9