Pushdown filters on coalesced outer join keys compared for equality under the join condition (https://redirect.github.com/duckdb/duckdb/pull/18169)
This PR implements filter pushdown on coalesced keys which are equal
under some full outer join condition (similarly to what has been
requested in https://github.com/prestodb/presto/issues/23110 and
https://github.com/trinodb/trino/issues/3722). Such filters are easy to
encounter, especially when relying on the `USING` syntax for the join,
which creates the coalesced expressions implicitly.

For example, the filter in the following query
```sql
SELECT id
FROM t1 FULL OUTER JOIN t2 USING (id)
WHERE id >= 2
```
can be safely pushed down into both sides of the join
```sql
SELECT id
FROM (SELECT id FROM t1 WHERE id >= 2)
FULL OUTER JOIN (SELECT id FROM t2 WHERE id >= 2)
USING (id)
```

More in detail:

Whenever a filter is of the form `P(coalesce(l, r))` or `P(coalesce(r,
l))` where `P` is some predicate that depends only on `coalesce(l, r)`
and there is a join condition of the form `l = r` where `l` and `r` are
join keys for the left and right table respectively, then we can
pushdown `P(l)` to the left table, `P(r)` to the right table, and remove
the original filter.

I think this optimisation is safe to do because
- for every row in the output of the unoptimised filtered outer join,
either
- `l IS NOT DISTINCT FROM r`, and so `P(coalesce(l, r)) = P(coalesce(r,
l)) = P(l) = P(r)` all hold and so the row belongs to the optimised
outer join of the two filtered tables also or
- `l IS DISTINCT FROM r` so that, under the outer join condition `l =
r`, we must have without loss of generality that `l IS NULL` and `r IS
NOT NULL`, and so `P(coalesce(l, r)) = P(coalesce(r, l)) = P(r)`, so
that `r` must come from a row in the right table that satisfies `P(r)`;
moreover, `l` must come from a NULL-extended row instead of a row in the
left table, so that pushing down `P(l)` to the left table would not
remove or alter the output row, and so the row belongs to the optimised
outer join of filtered table also.
- for every row in the output of the optimised outer join of the two
filtered tables, either
- `l IS NOT DISTINCT FROM r`, and so `P(l) = P(r) = P(coalesce(l, r)) =
P(coalesce(r, l))` and so the row belongs to the unoptimised filtered
outer join also or
- `l IS DISTINCT FROM r` so that, under the outer join condition `l =
r`, we must have without loss of generality that `l IS NULL` and `r IS
NOT NULL`, and so `P(r) = P(coalesce(l, r)) = P(coalesce(r, l))`, `l`
must come from a NULL-extended row, and no row exists in the left table
such that `l = r`, and so the output row belongs to the output of the
unoptimised filtered outer join also.

Since every row belongs to unoptimised filtered outer join if and only
if it belongs to the optimised outer join of the two filtered tables,
then this transformation should not change the output of the plan.

Upstream-commit: https://github.com/duckdb/duckdb/commit/84bd10b122d0ee4c05ed8c9ea1cc2946284287e5