Fix serialization backwards compatability for varargs functions (https://redirect.github.com/duckdb/duckdb/pull/18596)
In https://github.com/duckdb/duckdb/pull/18410 we added support for
creating function signatures with "template types" that are inferred and
resolved by the binder. This centralizes a bunch of binding logic,
removes the need for a lot of special cases and makes it much easier in
general to write functions that operate on nested types.

To make use of this, we changed the function signature of bunch of our
built in nested functions, particularly the `map_*` family of functions
that had particularly egregious and bug-prone binding code.
Unfortunately, for some deranged reason most of these were marked as
functions taking `varargs` when they really had no reason to. I was
under the impression that our existing function serialization code would
handle `varargs` properly and either 1. serialize the types that were
actually instantiated as the function argument types, or 2. at least
mark the function as taking `varargs`.

We do neither. functions with `varargs` are (in most cases) simply
serialized as having zero parameters, which causes an error on the
catalog lookup during deserialization when my fancy new type signatures
actually specify their parameters.

To remedy this, we now check during deserialization if the function
we're trying to deserialize have 0 parameters, but somehow have
expressions passed into it anyway. If thats the case we assume it must
be a varargs function. We then derive the types from the arguments when
looking up the function in the catalog.

I've also added a cast rule that allows implicit casts from `ANY` to
`TEMPLATE`. This should also only happen during function deserialization
between different versions of DuckDB.

Closes https://github.com/duckdblabs/duckdb-internal/issues/5581

Upstream-commit: https://github.com/duckdb/duckdb/commit/2ed9bf887f61a0ac226ab8c8f1164601d985d607