Add "Hash Zero" verification CI run (https://redirect.github.com/duckdb/duckdb/pull/18623)
This PR adds a build flag that causes DuckDB's hash function to always
return 0:
```
HASH_ZERO=1 make
```
This is used in the "Hash Zero" verification CI run, which is a
`relassert` build that runs the fast tests (and skips the compiled
ones).

With this flag active, everything in our hash tables collides with
everything, which stress-tests our collision resolution. This should
catch issues like https://github.com/duckdb/duckdb/issues/17891 before
they get merged into `main`.

Of course, this also makes our hash tables incredibly slow. Even some of
the fast tests become too slow to run, so I've also added a `require
no_hash_zero` that can be added to tests indicating that they should be
skipped when this flag is active.

This has not found any issues with our hash tables, so it seems like
https://github.com/duckdb/duckdb/pull/18374 has fixed the issue. It has
found a bug in `MapKeyCheck` in `value.cpp`, which did not properly
check if all keys in a `MAP` were unique (it only checked if the hashes
are unique), which this PR also fixes.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0c2b3a5572948fd257561450d38a94f8b51fc9d0