Change arrow() to export record batch reader (https://redirect.github.com/duckdb/duckdb/pull/18642)
This PR changes the `arrow()` method to export a record batch reader
instead of an arrow table.
I think this is desirable since users mostly want to interact with arrow
in a streaming fashion. This is especially important to avoid OOM of
naively calling `arrow()`.

We also have a dedicated function to explicitly return an arrow table
with `fetch_arrow_table()`, and a record batch reader with
`fetch_record_batch`.

We could also make `arrow()` generic by accepting the output format, but
I personally think that's an overkill.

Upstream-commit: https://github.com/duckdb/duckdb/commit/59b98176d6d1ec23a21dc5acce8f9fbf38e2cfe6