[Dev] Add script to create patch from changes in an extension repository (https://redirect.github.com/duckdb/duckdb/pull/18620)
This script should aid in creating patches for extensions

It verifies that the `GIT_TAG` in `duckdb` matches with the checked out
hash in the extension repository.
It then does a diff (ignoring submodules, since we only want changes to
the repository itself) to create the patch with the given name.

It detects when a patch by the same name already exists, which can be
overwritten with `--overwrite`, in case you're making incremental
changes to your patch, or you need to fix a bug etc..

It also tries to detect conflicts between existing patches, checking
that they alter the same line in the same file.
This still requires manual intervention, you likely have to merge the
two patches in that case.

Upstream-commit: https://github.com/duckdb/duckdb/commit/eeac806bb06d10c42916c81af6059b6b0e9fd18e