[Fix] Correctly handle table and index chunks in WAL replay buffering (https://redirect.github.com/duckdb/duckdb/pull/18700)
Fixes an issue where some code paths pass the entire table chunk to
`DataTable::AppendToIndexes`, while others only pass the "index chunk"
containing only indexed columns.

Now, we always pass both the `table_chunk` and the `index_chunk`, and a
mapping between them. In the absence of indexes, or if there are no
bound indexes, we skip most of the work.

I've also already added the changes of
https://github.com/duckdb/duckdb/pull/18683 here, as they reproduce
multiple bugs fixed by this PR. I'll properly merge ossivalis once this
PR is in. :)

Upstream-commit: https://github.com/duckdb/duckdb/commit/f6dfad0f1a9af2bb17203a3992dad54ebdc80dc0