Python-style positional/named arguments for macro's (https://redirect.github.com/duckdb/duckdb/pull/18684)
This PR implements Python-style positional/named arguments for macro's.

Currently, DuckDB treats macro arguments with and without a default
value differently:
```sql
create macro m(a, b := 'b') as a || b;
select m(a := 'a');
-- Macro function m(a, b := 'b') requires a single positional argument, but 0 positional arguments were provided.
select m('a', 'b');
-- Macro function m(a, b := 'b') requires a single positional argument, but 2 positional arguments were provided.
select m(b := 'b', a := 'a');
-- Macro function m(a, b := 'b') requires a single positional argument, but 0 positional arguments were provided.
```
Arguments without a default can only be used positionally, and arguments
with a default can only be used by name.

This PR removes this restriction, and accepts all three of the above:
```sql
create macro m(a, b := 'b') as a || b;
select m(a := 'a');
-- ab
select m('a', 'b');
-- ab
select m(b := 'b', a := 'a');
-- ab
```
Arguments without a default can now also be used by name, and arguments
with a default can now also be used by position. If all arguments are
passed by name, they can appear in any order.

I've also cleaned up a few error messages. For example:
```sql
create macro m(a, a) as a || a;
-- OLD: table "0_macro_parametersm" has duplicate column name "a"
-- NEW: Duplicate parameter 'a' in function definition
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/61f797f0e2d17d16430e7ca6f52ce720ebd156c3