Secrets: if serialization_type is not specified, assume it's a key value secret (https://redirect.github.com/duckdb/duckdb/pull/18743)
Currently the `serialization_type` is never serialized by any secrets,
as this was added later, and serializing it would break forwards
compatibility. All of our secrets are key/value secrets - yet we assume
on deserialization that the secret type is `CUSTOM` which requires us to
then load/have the extension in order to deserialize the secret. By
changing the default secret type to `KEY_VALUE_SECRET` we can
deserialize all of our secrets correctly. If, in the future, we need to
create/serialize a non-key-value secret, we can set `serialization_type`
appropriately.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f66c91138fff766b5adb88a4f08993160c22f53b