Add OS X notarization for DuckDB CLI and libduckdb.dylib (https://redirect.github.com/duckdb/duckdb/pull/18747)
This PR adds Apple
"[Notarization](https://developer.apple.com/documentation/security/notarizing-macos-software-before-distribution)"
for the DuckDB CLI tool `duckdb` and the dynamic library,
`libduckdb.dylib`.

Basically, binaries are signed and then sent to Apple for approval. OSX
will check with Apple on first launch of a binary if it has been
notarized and allow certain things if that's true.

Notarization of command-line-binaries is not exactly a priority for
Apple, so OSX will still complain if one downloads the .zip with the
CLI, and then double-clicks it in Finder. *However*, with this PR, OSX
will allow starting this binary *from Terminal*, which did not used to
work.

Upstream-commit: https://github.com/duckdb/duckdb/commit/29c5e895fccda360646a2e5a6c59af1b9aad8916