Fix radix partitioning with more than 10 bits (https://redirect.github.com/duckdb/duckdb/pull/18761)
The radix partitioning code seems to have a copy/paste error. When 11 or
12 bits were requested for partitioning, `RadixBitsSwitch` dispatches to
a partition with just 10 bits. I can't speak to the impact, as I just
found it while reading the code.

(Apologies for the drive-by action. I couldn't think of an issue
description that is better than the diff. Double apologies if this is
not a bug but intentional!)

Upstream-commit: https://github.com/duckdb/duckdb/commit/d1bc92caa6c2a7cbf32a270e0c0d54a47dde7a92