Fix index resolution when querying table with index via view (https://redirect.github.com/duckdb/duckdb/pull/18319)
Index scans would fail to match as in https://redirect.github.com/duckdb/duckdb/pull/17290 if the position of the
column used in the filter did not match that of the `index_expr`: the
ART `filter_match` assertion fails:


https://github.com/duckdb/duckdb/blob/0b83e5d2f68bc02dfefde74b846bd039f078affa/src/execution/index/art/art.cpp#L160-L161

It appears to work in the direct table query case (as shown in the
issue) because a projection puts the indexed column first (so the
binding matches).

I am unsure if this is the correct approach/place to make the change,
however rebinding the column ref to its correct position relative to the
input projection appears to fix the issue. Included is a test that fails
on v1.3.2.

Fixes https://redirect.github.com/duckdb/duckdb/pull/17290 

Upstream-commit: https://github.com/duckdb/duckdb/commit/ebe58746a0a9e93371639ad4c101ee4c54c2e9b6