Improve autocomplete suggestions (https://redirect.github.com/duckdb/duckdb/pull/18773)
Fixes https://github.com/duckdblabs/duckdb-internal/issues/5584

This PR improves the autocomplete in a couple of ways: 
1. Aliases of settings are now included in the suggestions (such as
`max_memory` as an alias for `memory_limit`).
2. Options registered by extensions are included in the suggestions. 
3. If the user input matches a substring of a given autocomplete
suggestion it is given a higher score. In the code I give a penalty
(increased score) to anything that isn't a substring to avoid scores
going below zero. For instance `SET directory` now gives as suggestions
`home_directory`, `temp_directory` and `secret_directory` as first three
suggestions.

Upstream-commit: https://github.com/duckdb/duckdb/commit/23da27f23b35c07d4825a4d1221e81063a91817f