Add support for non-aggregate window functions (https://redirect.github.com/duckdb/duckdb/pull/18788)
This PR adds support for implementing non-aggregate window functions...
as aggregates, which can only be used as windows. Wow, thats confusing.

DuckDB currently blurs the line between window function and aggregate
functions, which makes it possible to use any aggregate function as if
it were a window function (although, not the other way around). Because
the aggregate function interface is relatively simple, DuckDB will use
different execution strategies, so called `WindowAggregator`:s depending
on what window parameters are supplied, such frame bounds, ordering,
distinct etc. etc. in attempt to accelerate certain window
configurations without necessarily adding a bunch of complexity to the
aggregate interface.

If none of the specialized execution strategies can be used, the window
operator will first attempt to use the `CustomWindowAggregator` which
uses the optional `window` and `window_init` callbacks that an aggregate
function can supply, before falling back back to using the "naive"
executor which also uses the standard aggregate interface, or

However, this becomes problematic if you want to define a window
function **which can only be used as a window function**. Simply
throwing an exception in the normal aggregate callbacks is not
sufficient because the window operator can still decide to use a
different strategy when windowing which still uses the aggregate
callbacks, instead of always instantiating the `CustomWindowAggregate`
which uses the window callbacks.

Ideally we do have a separate `WindowFunction` class/catalog entity to
make this distinction, but for now I've attempted to support my use case
by instead allowing the normal aggregate callbacks to be set to
`nullptr`, and inspecting whether they are present during binding of
aggregates/window expressions as well as choosing the window execution
strategy.

Upstream-commit: https://github.com/duckdb/duckdb/commit/a35bb84e8d48f7102da9b811ea20ab74d1c7a0d9