Add callback to get a list of copy options, use this to provide suggestions and to erase options from import that are only used during exporting (https://redirect.github.com/duckdb/duckdb/pull/18812)
Fixes https://github.com/duckdb/duckdb/issues/17858

This PR adds a new callback to the `CopyFunction` - `copy_options`. This
is used to provide a list of options provided by the copy statement, as
well as (optionally) their types and whether or not they can be used
when reading, when writing, or for both.

This allows for a few things:
* When running `EXPORT DATABASE`, we know if an option is for writing
only, or if we need to mirror it in the `COPY` statements we generate
* When parsing an unknown option, we can now give suggestions:

```sql
copy (select 42) to 'file.csv' (partition_b x);
```
```
Not implemented Error:
Unrecognized option "partition_b" for csv

Candidate options: "partition_by", "separator", "compression", "write_partition_columns", "prefix"
```

* When using an option that is used for writing but can only be used for
reading, or vice versa, we can now inform the user of this explicitly:

```sql
copy (select 42) to 'file.csv' (encoding 'utf-16');
Invalid Input Error:
Option "encoding" is not supported for writing - only for reading
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/da1eb68203a47dff51d30bbf7ca76fbfcfe9597a