Fix: Preserve database configuration flags for tab completion in DuckDB shell (https://redirect.github.com/duckdb/duckdb/pull/18482)
When using tab completion in the DuckDB shell, the completion system
creates a new database connection without preserving the some important
configuration flags. This causes issues when extensions themselves load
other extensions as part of tab-completion.

When a database is opened in the DuckDB shell and a tab-completion is
requested, the database config that is used does not pass the flag for
`allow_unsigned_extensions`, in actuality no config is passed. This PR
persists the open flags that are set for the database so that it can be
used as part of `linenoise_completions`.

This PR modifies the shell to persist the database configuration flags
that were used when opening the database, ensuring they are available to
the `linenoise_completions` function.

I encountered this bug by:

I have an extension that in its load function attempts to load other
extensions, some of which are unsigned.

If this extension is loaded by the `linenoise_completion` function,
which pretty easy (just start the shell and press tab), the value for
`allow_unsigned_extensions` is false even though the DuckDB shell was
started with the `-unsigned` CLI argument.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c1db86fa63702b94c0bff59e4fa18f0ee89b48fc