Ensure a WAL file matches the DB file and checkpoint iteration (https://redirect.github.com/duckdb/duckdb/pull/18823)
When using the latest storage, we now add metadata to the WAL to ensure
it matches the DB file. To do so, the DB files now have unique
identifiers, and the WAL files contain these identifiers as well as the
current checkpoint iteration of the DB file.

Close https://github.com/duckdblabs/duckdb-internal/issues/5657

Upstream-commit: https://github.com/duckdb/duckdb/commit/940c8aa5ef4708fd621c338046de4ec16ec2fef0