Add support for reading/writing native parquet geometry types (https://redirect.github.com/duckdb/duckdb/pull/18832)
This PR extends DuckDBs parquet-geometry support by writing spatial
geometries as the new native parquet `GEOMETRY` logical type.

Parquet recently standardized support for `GEOMETRY` and `GEOGRAPHY` as
logical types into the parquet spec itself. This is in theory mostly
compatible with the existing "GeoParquet" extension format, at least to
the extent that DuckDB supports it so far. While the as-of-yet
unreleased "GeoParquet 2.0" is supposed to reconcile and unify the two
specifications, the changes in this PR favors writing files following
the native spec, while also writing the required GeoParquet metadata to
the best of our ability, as the new native geometry spec is basically a
superset of DuckDBs existing geoparquet support. In practice, this means
that exporting a dataset containing a spatial `GEOMETRY` type will
produce a parquet file that is both a valid GeoParquet file and a valid
"native geometry" parquet file. As the new native types are "logical
types", older parquet readers that don't support the native types yet
should still be able to read these columns as their underlying physical
type, and if they support GeoParquet they should still be able to
interpret these as geometries as the actual geometry encoding to binary
(ISO WKB) is the same.

In the few cases where the specs differ, e.g. when writing geometries
with "M" coordinates (which GeoParquet does not support), we omit the
geoparquet metadata for those columns.

We also write the new geospatial stats. We can't make use of them
ourselves yet when reading, but it's possible to inspect the bounding
box and geometry type stats through the `parquet_stats` function . By
utilizing the native parquet geometry stats we can clean up a lot of the
existing "special case" geoparquet code by reusing the same statistics
collection codepaths. I've also implemented the WKB stats computation
code into the extension, avoiding the need to call into spatial when
writing.

This PR also adds support for returning the geospatial statistics when
using the `RETURN_STATS` option of the `COPY` command, opening the path
for writing geometries to different data lakes.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0269779f0dd5f905261f9c547a17c1fca20d95d4