Centralize attached database paths in a DatabaseFilePathManager which is shared across databases created through the same DBInstanceCache (https://redirect.github.com/duckdb/duckdb/pull/18857)
Fixes https://github.com/duckdb/duckdb/issues/14421

On POSIX systems (Linux / MacOS) file-based locks are granted on the
process level. As a result, we cannot rely on them to ensure safe
locking within the same process. Instead, we need to have a central
location for co-ordinating which files have already been opened.

Currently we have two ways of doing this:
* The `DBInstanceCache` ensures top-level connections (e.g.
`duckdb.connect('file.db')` in Python) have a shared list of attached
files, and when two connections to the same database file are made, they
are routed to the same top-level connection.
* For attached databases, we have a list of database paths in the
`DatabaseManager` (`db_paths_to_name`) - we use this to ensure within a
single database the same database file is not attached multiple times.

This, however, leaves one issue: the list of attached databases is not
shared across separate top-level conections. As a result, we can open
multiple top-level connections, and then attach the same database file
in those top-level connections. This can lead to concurrent writes to
the same database file, which can lead to e.g. data corruption as the
regular locking invariants are violated.

This PR centralizes the list of attached databases in the
`DatabaseFilePathManager` class. This list can be shared across
top-level connections. Any connection created through the
`DBInstanceCache` gets a reference to the shared list of attached
databases. This, in turn, prevents the same database from being attached
in different top-level connections.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b505737f9816275a57bd2c6e2baa19685f25c801